Add 7/30 day pageview reports; genericize descriptions
If you're not running multiple domains (which I'm going to assume
is most organizations), the 7/30 day page views report is more
useful than the 7/30 day domains report.

The descriptions all says ".gov" and "digital analytics program"
for every report.  No need to be that specific, especially if
we want others to use the analytics-reporter.  Remove this
identifying information.