Fix issues with paths in cloud.gov
There were some issues with the way paths were managed in cloud.gov.
This commit has some fixes for those.

The `$HOME` directory was expected to be `/home/vcap`, but it looks like
that is `/home/vcap/app` when the app is started. This commit changes
the `daily.sh`, `hourly.sh`, and `realtime.sh` to use the correct
directory in response to that.

This commit also changes the line count output in `cron.js` so that it
uses the correct path when looking for the log files. This was causing
the app to crash beforehand.

Finally, this commit updates `cron.js` so that is uses the correct
relative paths when looking for the scripts that generate the reports.