Make data paths consistent with slugs for analytics.usa.gov
This commit makes the paths where the data is accessed consistent with the names of the slugs for the given agency on analytics.usa.gov. This has the advantange of not requiring us to manage separate data and agency paths.

This is in reference to #172 to complete the following task:

> Rename directory slugs for agency data to match slugs used in analytics.usa.gov (e.g. social-security-administration instead of ssa).