Output results from scripts in cron.js
This commit outputs the results from running `realtime.sh`, `hourly.sh`, and `daily.sh` in `cron.js`.

It does this by replacing `exec` with `spawn`. `spawn` returns an event emitter which, unlike `exec`, allows us to monitor the output in real time. In addition, the logs will have been outputted before the script exits. This ensures the integrity of our logs if the reporter crashes or is stopped with one of these scripts running.

ref #172