Add New Relic
This commit adds New Relic for the analytics reporter. It is configured to use an App Name and License Key provided by a cloud.gov user provided service named "analytics-reporter-new-relic".

Additionally, I added a `NODE_ENV` environment variable in production to enable the activation of New Relic in production only.

ref 18f/analytics.usa.gov#428