Update Travis config to use node 7
This commit updates the Travis config to use node v7.

The package.json was updated to include node v7 in a previous commit, and some ES6 syntax was added. The Travis config was not updated, however, causing the tests to fail.

This commit updates the config so Travis will know to use the more recent version of node.