Move GA result processing into external function
This commit removes `Analytics.process` and replaces it with a separate function that lives in its own module. This allows for some cleanup of the analytics module by modularizing and removing the logic necessary to transform raw GA results into our format.

This commit also sets up some basic tests for the transform function.