Break tests for calculate totals into a separate file
This commit removes the tests that verify that totals are caclulate
accurately from the process-ga-data tests and moves them into a new file
named calculte-totals.test.js. This matches the pattern set forth by the
tested code.

This commit also adds a test to process-ga-data to verify that it calls
out to the calculateTotals function.