Add a module that loads GA credentials on a rotation
This commit adds a function that provides secret keys and report emails on rotation. This is necessary because we have to maintain several setso f DAP credentials to stay under Google's rate limits.

This commit lets the analytics reporter call a function to get a set of creds which are provided on rotation. This means an even dispersment of creds amongst API calls.

The creds are loaded by looking for environment variables with the name `ANALYTICS_CREDENTIALS`. This is expected to be either a JSON encoded object with the following schema:

```
{
  "email": "ga_email@example.com",
  "key": "a private key"
}
```

...or a JSON encoded array of such objects. If an array is given, every call will return the next object in the array, going back to the first when it reaches the last.

Ref #183