Use travis for continuous deploy
This commit deploys from travis using the cloud foundry provider. #193
moves all of the things the app needs to deploy into env variables, so
credentials do not need to be in a developers machine. This makes it
possible to deploy from CI because we won't need to save these files on
the machine deploying.