Remove dap-1 and dap-2 env files
This commit cleans up the deploy folder so that the daily.sh, hourly.sh,
and realtime.sh files will stop calling out to dap-1.env and dap-2.env.
Now that we are setting the credentials provided by those files in the
environment, we no longer need to load them before fetching reports.
Moreover, when deploying from CI, attempting to load those files will
result in an error since they do not exist on the CI server.