Overrite rows if visits or users change
Some of the reports have values that change through out the day. For example, the `today` report will have a data point that looks like this:

```json
{
  "date": "2017-03-08",
  "hour": "15",
  "visits": "0"
},
```

...until 3:00 PM, when the `visits` prop will start changing. Prior to this commit, the `writeToDatabase` function did a strict comparison of rows to determine if a row already existed. This row would have been inserted multiple times as the `visits` prop changed throughout the day.

This commit fixes the problem described above by building a query that does not include `data --> 'visits'` and `data --> 'users'` when looking for existing rows. If an existing row is found with this new query, the function looks to see if those props have changed. If they have, it updates the existing row.