Consolidate GA logic into GoogleAnalyticsClient
Prior to this commit, `analytics` module was calling out to the various GA related modules to build, authorize, and send queries for data. This commit takes all of the logic necessary to send a GA query and loads it into the `GoogleAnalyticsClient` module. This allows for all of the GA logic to be invoked with a simple `GoogleAnalyticsClient.fetchData(report)` call.