Temporarily disable postgres publisher
The database publisher is causing the reporter to run slowly as the
database swells. We've discussed the need for a new solution to manage
the data, but have not decided on one yet. In the meantime, the database
is collecting more and more data, meaning the reporter runs slower and
slower on reads and writes.

This commit temporarily disables the PostgresPublisher until such time
as we find a better solution.