Add Docker Compose configuration
This commit adds a Docker Compose configuration that can be used to run the Analytics reporter as if it were running on cloud.gov. This PR also includes some scripts to aid in running the reporter in Docker, as well as some changes to the scripts in `deploy/` to make it a little more environment agnostic.