Add postgres db to docker compose
This commit adds a postgres db to the docker compose configuration so the analytics-reporter can write to a db when running in docker compose.

A tradeoff is that the migrations had to be copied into this app from analytics-reporter-api. Hopefully, if we integrate analytics-reporter-api into our docker-compose configuration, we can remove the migrations from this repo and use the ones in the API repo.