Fix gitignore to be compatible with dockerignore
There was some syntax in the gitignore that was incompatible with the
dockerignore. This commit updates that syntax so they are in sync.

This commit also adds the `.git` dir to the gitignore so that any
secrets that may be in a working branch in the git repo won't be leaked.