Add separate API report that runs daily
This commit adds a new serious of reports in a report file named `api.json`. These are reports that are meant to add data to the postgres database that backs the API. This commit configures the reporter so that only the results from reports in `api.json` are written to the database. Additionally, `api.json` contains only daily reports so it only needs to be run every 24 hours.

This advantage of this change is it allows us to manage the reports that feed analytics.usa.gov and the reports that feed to API seperatly so we can manage the data going into one without having to worry about side effects in the other. Additionally, it gives us finer grained control over data written to the API.

Additionally, this relieves the reporter from having to process a lot of the redundant data in the standard, more far-reaching reports (e.g. looking at data that is up to 90 days old and has already be inserted).

The thresholds on these reports can also be managed separatly from the standard reports which allows for more aggressive downsampling.