Don't set timeout for daily reports to more than 24 hours
There was an issue with the daily reports. They are supposed to run every morning at 10 AM UTC which translates to 5 AM or 6 AM depending on DST.

To calculate when they run, we increment the day in the current date, set the time to 10 AM UTC, and find the interval between then an now. This create a problem when the reporter is started late at night. 2 AM UTC is 10 PM the previous day in ET. This means that if the reporter is run at 10 PM ET on 2017-01-01, the date in UTC is 2017-01-02, so when the day is incremented, it is set to 2017-01-03. This means that the reporter will not run the next morning (morning of 2017-01-02), waiting a full day before running again.

This commit fixes the problem, by finding the modulo of the interval and 24 hours. This will prevent the interval from ever going over 24 hours, while still using the offset between the current data and 10 AM ET on the next day.