Fix issue where daily.sh was exiting early.
The `daily.sh` script was misbehaving. It exits right around the time it started processing HUD data. After a little investigation, I realized this was because it was receiving a SIGTERM. This was because we were still using `child_process.exec` instead of `child_process.spawn`, which implement a similar API, but behave slightly differently.

`child_process.exec` has a max buffer size, meaning that whenever the the output exceeds a certain length, it exits and returns a string that says `Max buffer size exceeded`. But, since we were looking at the `stdout`, `stdin`, and `exit` events instead of the callback, we never saw this. As a result, the buffer size was being exceeded and the process was exitting without us being aware of why.

This commit changes require so we are using the intended function, and also makes some other slight changes to use the `child_process.spawn` API.

closes #18f/analytics.usa.gov#451