Filter NCBI hostnames from the top 20 realtime pages
The NCBI DAP implementation aggregates their pages. This means that requests for a bunch of pages are rolled into one, and that one page shoots to the top of the list. This commit filters NCBI so those pages don't pollute the top 20 since they aren't actually requests for a single page.