Create new api endpoint for longer/seasonal session data This endpoint would be used for a new barchart on analytics.usa.gov: https://github.com/18F/analytics.usa.gov
If we want to change the "Visits Today" barchart into something more seasonal, we may need to generate a new report since there isn't an equivalent report in the list.
This query is modeled off USA.gov's Last 30 Days endpoint:
https://app-usa-modeast-prod-a01239f-ecas.s3.amazonaws.com/analytics/raw-data/traffic-sources-30-days.json

We can tweak the number of days to last 90, last 365, or whatever length of days we decide.